+++ jest.setup.js
@@ -1,10 +1,7 @@
-/* eslint-env node */
-'use strict'; // eslint-disable-line
-
-require('@testing-library/jest-dom');
+import '@testing-library/jest-dom';
 
 // Node has an Intl object, but doesn't ship with any locale information, so we
 // need to patch parts of it to work as it does in the browser.
-require('@formatjs/intl-locale/polyfill');
-require('@formatjs/intl-numberformat/polyfill');
-require('@formatjs/intl-numberformat/locale-data/en-NZ');
+import '@formatjs/intl-locale/polyfill';
+import '@formatjs/intl-numberformat/polyfill';
+import '@formatjs/intl-numberformat/locale-data/en-NZ';