+++ source/MessageFormatterProvider.js
@@ -1,12 +1,12 @@
-/* 
+/*
  * Copyright 2020, Emanuel Rabina (http://www.ultraq.net.nz/)
- * 
+ *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- * 
+ *
  *     http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -18,11 +18,11 @@ import MessageFormatterContext from './MessageFormatterContext.js';
 
 import {MessageFormatter} from '@ultraq/icu-message-formatter';
 import PropTypes from 'prop-types';
-import React, {memo, ReactNode, useEffect} from 'react';
+import React, {memo, useEffect} from 'react';
 
 /**
  * @typedef {object} MessageFormatterProviderProps
- * @property {ReactNode} children
+ * @property {import('react').ReactNode} children
  * @property {MessageFormatter} formatter
  * @property {Record<string,string>} [messages]
  * @property {import('./MessageFormatterContext').MessageResolver | null} [messageResolver]