+++ source/withMessageFormatter.js
@@ -1,12 +1,12 @@
-/* 
+/*
  * Copyright 2020, Emanuel Rabina (http://www.ultraq.net.nz/)
- * 
+ *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- * 
+ *
  *     http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -16,7 +16,7 @@
 
 import MessageFormatterContext from './MessageFormatterContext.js';
 
-import React, {ComponentType, FunctionComponent} from 'react';
+import React from 'react';
 
 /**
  * A higher-order component function that applies the context objects,
@@ -25,8 +25,8 @@ import React, {ComponentType, FunctionComponent} from 'react';
  *
  * @author Emanuel Rabina
  * @template P
- * @param {ComponentType<P>} Component
- * @return {FunctionComponent<P>}
+ * @param {import('react').ComponentType<P>} Component
+ * @return {import('react').FunctionComponent<P>}
  */
 export default function withMessageFormatter(Component) {
 	const wrappedComponent = (props) => (